+++ packages/ricos-editor/web/src/toolbars/FloatingPluginMenu/PluginMenuButton.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { ToolbarButton } from 'wix-rich-content-editor-common';
+
+const PluginMenuButton = ({ Icon, label, onClick, t, tooltipText }) => {
+  const pluginButton = (
+    <button onClick={onClick}>
+      <Icon />
+      <div>{t(label)}</div>
+    </button>
+  );
+
+  return <ToolbarButton tooltipText={tooltipText} button={pluginButton} />;
+};
+
+export default PluginMenuButton;